server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.5.0
google-genai>=0.3.0
pymupdf>=1.23.0
python-docx>=1.1.0
python-multipart>=0.0.6
